Fix getsockopt syscall on s390x (#19)
The GetsockoptIPv4OriginalDst function uses a pointer to int as an argument for the getsockopt syscall. On 64-bit systems, the int type will be 64-bit, but getsockopt requires a pointer to a 32-bit integer.

On little-endian architectures (amd64, arm64, etc.), a pointer to a 64-bit integer is the same as a pointer to a 32-bit integer if the integer is small enough. On big-endian architectures (e.g., s390x), this is different.

Update the GetsockoptIPv4OriginalDst function to use the correct type for the size argument and apply some small refactoring to this function.